[Monorail] Fixed SQL generated for custom field search.
Add the case for generating SQL for a query condition on a date-type field.
Also, change the SQL generation logic to include conditions in the
JOIN clause and use the WHERE clause only to check if some rows (or none)
were successfully joined.  That is the pattern followed by other condition
processing functions and is needed to handle multi-valued fields.

BUG=monorail:1519

Change-Id: I662b416ddb6dcbf36c1d0ccac28815bc15a79378
Reviewed-on: https://chromium-review.googlesource.com/394286
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>