[Issue wizard] Add better tracking of wizard form and label interop issues.
The better Via-Wizard label is just a one-line change.

The Hotlist-Interop change is a simple function that is tied to just
the <select> element on the API form.  Form elements with name="labelN"
are copied to the form that is posted to issue.do, and label2 is
previously unused.

BUG=1765

Change-Id: I99a6e15143af522a634e8b787065012171573999
Reviewed-on: https://chromium-review.googlesource.com/395640
Reviewed-by: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>