Generalize search error messaging
Right now, we only support showing the user one kind of search error:
"Some of the backends failed. Sorry.". We need a more general framework
for displaying errors to the user, for everything from "some backends
failed and we don't know why" to "you entered an invalid query, here's
the bad token". This is a first step in that direction.

BUG=monorail:1839

Change-Id: Ib2a6415ec3a7c8fedc7aa4eb81ff01c579bb29cd
Reviewed-on: https://chromium-review.googlesource.com/397478
Commit-Queue: Aaron Gable <agable@chromium.org>
Reviewed-by: Jason Robbins <jrobbins@chromium.org>