Allow colons in the value of a key:value search query
We have code which supports processing queries like "blockedon:skia:123",
but the first-pass query parsing code sees the two colons and defaults
to treating it as a fulltext search instead. This CL updates the regex
used to parse these key:value tokens to support including some more
special characters (in particular, colon) in the value, as long as they
are not the *first* character of the value.

BUG=monorail:1839

Change-Id: I2603b231f79ab5fc040163d8277d7f72d37bfe94
Reviewed-on: https://chromium-review.googlesource.com/397818
Commit-Queue: Aaron Gable <agable@chromium.org>
Reviewed-by: Jason Robbins <jrobbins@chromium.org>