kitchen: use "./workdir" as default workdir
by default kitchen uses a directory under $TEMPDIR if workdir is
specified. This is not practical on our fleet. Use "./workdir" instead.

Also do not delete it after a run. It is practical for debugging and on
swarming, workdir will by inside the rundir which is deleted each time.

R=vadimsh@chromium.org
BUG=

Change-Id: I0817647010da6f6fbbd857f7ec453af13033509a
Reviewed-on: https://chromium-review.googlesource.com/404531
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>