[som] prep to generate alerts feed from pubsub persistent store
This doesn't generate the actual AlertsJSON object yet, but it does add
the necessary code to go from /pubsubalerts/:tree to getting the 
masters for that tree and querying the datastore for alerts on each 
tree.

This includes a workaround for reading gatekeeper_trees.json from 
luci-config: It reads the file from gitiles and cahces it in memcache.
Not ideal since it can't read internal tree configs in this state.

BUG=655286

Change-Id: I1b56bf570e5cb67820a4efe0a083bf88184e6282
Reviewed-on: https://chromium-review.googlesource.com/405781
Commit-Queue: Sean McCullough <seanmccullough@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>