[Monorail] Avoid 500 on [commentby:email1,email2].
I noticed that that type of query was failing our internal safety check
and causing 500s.  The main change is a new part of a RE in JOIN_RE_LIST.
I also added a new unit test and lots of calls to our _IsValid.*() functions
in the unit tests.

BUG=monorail:1839

Change-Id: Ie1813474f1c0faa439dce3c6edb8b760d99ede29
Reviewed-on: https://chromium-review.googlesource.com/412364
Reviewed-by: Aaron Gable <agable@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>