[Monorail] Offer AllLabels column in the web UI.
This was previously automatically added to CSV output and not offered
in the UI.  Now, we can offer it in the "..." menu and users who know
to look there can add it as a column in the web UI.

Most of the code change is just moving functions around in the file.

Also, I add a <wbr> element in multi-valued table cells between each
value to give the browser a hit about where to break lines.  Since
this works so well in combination with NOWRAP, I made that the default
for all cell types except Summary.

BUG=monorail:1932

Change-Id: I5600ad13944444ab5d04577d38a6ea5f30a7a999
Reviewed-on: https://chromium-review.googlesource.com/412976
Reviewed-by: Aaron Gable <agable@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>