[Monorail] Fine-tune the display of the color block help cue.
This is a follow-up to https://chromium-review.googlesource.com/c/414306/.
The problem being fixed is that when the issue has owner_id == 0 or
derived_owner_id == 0, those are still created as UserView objects and
they have message "Never visited", so they trigger the help cue even
when there is no color block displayed on the page.  It is better to
only show this help cue when it is relevant so that people understand
the feature rather than dismissing the help cue early and then forgetting
that they ever saw it.

BUG=monorail:305

Change-Id: I95f35488c9c8eb30c1cf3f7472daf13f1c075425
Reviewed-on: https://chromium-review.googlesource.com/414986
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>