[Monorail] Allow sorting by ownerlastvisit.
In this CL:
+ Generate SQL JOIN and ORDER BY terms for sorting by ownerlastvisit.
+ Implement sorting key accessor for use with sorting.py.
+ Generalize sorting.py key post-processing.

The way we sort in RAM is to pass "accessor" functions to the standard
python sorted(issues, key=...) function.  We were already constructing
some accessor functions that get a field from the Issue and do a post-
processing step to look up the user_view and get the email.  This CL
generalizes that to apply a post-processor function instead of always
assuming we want the email address string.

BUG=monorail:305

Change-Id: I114fb85f15a45cf1f66fc65bf43beb54fc2328f2
Reviewed-on: https://chromium-review.googlesource.com/414530
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>