ts_mon: Continue flushing metrics beyond serialization error.
Before this CL, If flush() encounters an error while serializing a
metric, flush immediately aborts. This means that
- A subset of the metrics get flushed (whatever batches were already
  sent)
- ts_mon never sends the remaining metrics till the offending metric is
  reset externally / the ts_mon flushing thread restarts.

This CL hardens flush() to continue flushing the rest of the metrics
beyond any corrupt metrics encountered. flush() continues to raise an
exception, so the failure mode is unchanged.

This is the first of a two part change. This only changes the code path
for serialization to old proto type.

BUG=chromium:669669
TEST=(new) unittest fails before change, passes afterwards.

Change-Id: I2c158144cc6fd1f4a7b40ca7a625c7ff2400bfcc
Reviewed-on: https://chromium-review.googlesource.com/415447
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Sergey Berezin <sergeyberezin@chromium.org>