ts_mon: Continue flushing metrics beyond serialization error [2/2]
Before this CL, If flush() encounters an error while serializing a
metric, flush immediately aborts. This means that
- A subset of the metrics get flushed (whatever batches were already
  sent)
- ts_mon never sends the remaining metrics till the offending metric is
  reset externally / the ts_mon flushing thread restarts.

This CL hardens flush() to continue flushing the rest of the metrics
beyond any corrupt metrics encountered. flush() continues to raise an
exception, so the failure mode is unchanged.

This is the second of a two part change. This changes the code path for
serialization to new proto type.

BUG=chromium:669669
TEST=(new) unittest fails before change, passes afterwards.

Change-Id: I2e1bb1be7e63c643922eb3548ef34bc02eb053ee
Reviewed-on: https://chromium-review.googlesource.com/415584
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Sergey Berezin <sergeyberezin@chromium.org>