Adding memoized functions
Memoization is necessary when training loglinear models to avoid recomputing features over and over. Recomputing features affects not just the constant factors but also the asymptotic complexity (in a way that's already observable for small ``N``). The memoization is factored out into a class because doing so greatly reduces the code complexity over doing it manually in all the places where we need it.

BUG=
TBR=stgao@chromium.org

Review-Url: https://codereview.chromium.org/2548603003