Implementing loglinear models, for CL classification
This CL adds an implementation of loglinear probability models, and translates the ./crash/scorers into feature functions for use with those models. It does not, however, put LLMs in place for being used as the primary algorithm for CL classification just yet. That should be done in a separate CL accompanied by delta testing to ensure it is as effective as we'd like.

BUG=
TBR=stgao@chromium.org

Review-Url: https://codereview.chromium.org/2517383005