[Monorail] Escape CSV formulas.
Prevent any hacker from attempting to trick another user into loading
one of our CSV files into a spreadsheet app which would then interpret
one of the cells as a formula which could have a side-effect.

This also eliminates TableCellSummaryCSV because the only purpose of that
class was to do escaping, and now escaping is done in the servlet for all
values in all cells.

BUG=monorail:1996

Change-Id: I5353e3b5bb8b275c40a66e30b923332938a20b39
Reviewed-on: https://chromium-review.googlesource.com/418450
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>