[Monorail] Only log the name of the API client that actually called us.
Previous to this change, we log the email address of every whitelisted
API client on every request.  That makes it useless to search the
logs for the address of a given client.  Now, we only log that info
to help with debugging a permission exception.

BUG=monorail:1998

Change-Id: I45345820ba35ba53ee576edc39303167420ceff5
Reviewed-on: https://chromium-review.googlesource.com/418481
Reviewed-by: Aaron Gable <agable@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>