Just commit gitlinks, don't mirror the whole source tree.
This changes the destination repo to be more of an "overlay" repository just
containing gitlinks, rather than a mirror of the repository with gitlinks added
in a commit on top.  This means we can stop rewriting history on each commit,
and the codesearch crawler can use diffs between revisions of the overlay
repository to only crawl the changed revisions inside subrepositories.

BUG=

Change-Id: I6a4a63ef59faf3180be1e78b6c79019e9d72a75f
Reviewed-on: https://chromium-review.googlesource.com/413787
Commit-Queue: Dave Sansome <dsansome@chromium.org>
Reviewed-by: Emma Söderberg <emso@chromium.org>