Continuing hotlist flipper.
The hotlist flipper is preserved when making changes (adding/deleting comments)
in each issue. 
The sort order of the hotlist is also being transferred to the details page for the flipper
to use.
TODO left: clicking on the row vs clicking exactly on the issue_id in the row still does not
take me to the same flipper.

staging url: https://7180-7f2a1c6-tainted-jojwang-dot-monorail-staging.appspot.com

BUG=monorail:1871

Change-Id: I5140f0d88355e444fc0248e5c94282794e1e47d4
Reviewed-on: https://chromium-review.googlesource.com/419741
Reviewed-by: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>
Commit-Queue: Joanna Wang <jojwang@google.com>