AD: Use test suite name instead of step name
This change causes alerts dispatcher to generate alerts for tests using
the test suite name, rather than the raw step name, because the raw step
name often has extraneous information in it. For now, we're doing some manual
parsing, but eventually this should be done with some sort of structured data.

Also removes parallelizing from the convey tests. It messes with the output in
the goconvey tool.

Change-Id: If0b32c2f9d92d2e7217a57b1083577b6b33a68b0
Reviewed-on: https://chromium-review.googlesource.com/419708
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>