[Monorail] Display initial owner availability info.
This only shows availability info for the existing owner of the issue,
or the initial owner on the issue entry form.  It does not yet update as
the signed in user edits the issue owner, but this CL does establish an
HTML element ID for the <span> that will need to be updated when we do
dynamic display of availability info.

BUG=monorail:2027

Change-Id: Ida5c35da14dfe21d0b493468a60c0fa242502514
Reviewed-on: https://chromium-review.googlesource.com/422529
Reviewed-by: Aaron Gable <agable@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>