[Predator] Add flag manager and flag manager test.
Flag manager is a class to track all the flags during the stacktrace parsing. During the stacktrace parsing, some flags are needed to determine priorities of callstack or whether a frame of a stack should be filtered or not. For example, if 'fatal_java_exception' flag is on, the first java callstack we detected with this flag should have high priority.

BUG=

Review-Url: https://codereview.chromium.org/2593093003