[Monorail] When viewing a hotlist, show username first in main tabs.
This makes the ordering of main tabs in across the top of the page
consistent between the user profile page, the user list of hostlists,
and the details of any one hotlist.  Also, I think people read
left-to-right and assume containment in that order, so putting
the user name first make it clear that the hotlist is in that
user's URL space.

BUG=monorail:50

Change-Id: I67551876d047908962d7231c10a97d5038a9d573
Reviewed-on: https://chromium-review.googlesource.com/424914
Reviewed-by: Joanna Wang <jojwang@google.com>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>