Adding CrashReportWithDependencies
This new class extends CrashReport with dependency and dep_rolls information. This information was previously computed on an ad-hoc basis, but now this class helps encapsulate the fact that this is feature- and suspect-independent metadata used by individual features in the loglinear CL-classifier. In addition, we document and more fully specify what exactly we're doing when we compute this metadata (i.e., exactly which dep_paths we keep information about and why).

This same class could be used to clean up the old scorer-based CL-classifier too, but we do not do that here since it would make the CL overly large (and the scorer-based classifier is due to be deprecated anyways).

In addition, since this class is only used by the CL-classifier (so far), we convert from CrashReport to CrashReportWithDependencies within LogLinearChangelistClassifier.__call__ rather than in Predator.FindCulprit. This helps reduce coupling / maintain modularity. If in the future it turns out we also want to pass this annotated crash report to the other classifiers, then we can move the conversion to predator.py.

This is part 1 of https://codereview.chromium.org/2618553005/

BUG=
TBR=stgao@chromium.org

Review-Url: https://codereview.chromium.org/2617743003