AD: Add links from buildbot steps
This CL finds any links which are present in the buildbot code (added by
the recipe), and adds that to the alerts json. This will help provide additional
information for sheriffs; recipes can link screenshots, logs, traces, etc..., and
this information will show up on sheriff-o-matic.

There's the potential for some issues, like really long link names, but I think we
can cross that bridge when we get to it. There's also some security risk, since
we are allowing recipes to put arbitrary links on sheriff-o-matic, but if an untrusted
user can change executed code in recipes, we're already in trouble.

Change-Id: I4324d969409390dd5d5ee52ab1a54f234906421b
Reviewed-on: https://chromium-review.googlesource.com/424895
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>