Actually compare TestFile.Builder field with builder stored in JSON
Although the previous version of the code should be doing the same because we
are retrieving TestFile entity based on upload params (p.Builder), but it would
be more semantically correct to check TestFile.Builder field directly. This
would prevent us from deleting valid entities if there is a bug in p.Query that
would cause it to return arbitrary entities.

BUG=678026
TBR=tandrii@chromium.org

Change-Id: I69305fa81bc1a26a733fd640781c120c9b72613a
Reviewed-on: https://chromium-review.googlesource.com/429990
Commit-Queue: Sergiy Byelozyorov <sergiyb@chromium.org>
Reviewed-by: Sergiy Byelozyorov <sergiyb@chromium.org>