[Monorail] Make multi-valued table cells word-wrap on other browsers.
Previously I used
<td class="nowrap">
  val,<wbr>val,<wbr>...
</td>
That worked fine in chrome and it avoided the need to have another
level of DOM elements in the common case of only one value in a table
cell.  However, FF and IE/Edge do not seem to handle that.  So, now
I am going with more explicit markup that looks like:
<td class="nowrapspan">
 <span>val,</span>
 <span>val,</span>
 ...
</td>

BUG=monorail:2093

Change-Id: I4ce168023a950aa1960307320134da03d042d703
Reviewed-on: https://chromium-review.googlesource.com/430271
Reviewed-by: Joanna Wang <jojwang@google.com>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>