[Monorail] Deflake comparison of XSRF tokens by mocking time.time().
I happened to see a test failure in this test where the XSRF tokens are
compared.  The cause of the failure was that generating the two tokens
happened a split-second apart, which sometimes sees different values for
int(time.time()).  Mocking time.time() to always return the same value
should eliminate that possibility.

BUG=monorail:50

Change-Id: I841eca256f33c8a009857d109c4596b04e7b1585
Reviewed-on: https://chromium-review.googlesource.com/431140
Reviewed-by: Joanna Wang <jojwang@google.com>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>