swarmbucket: fail builds for non-exitent builders
Historically swarmbucket ignores builds for undefined builders. This was
done so we can migrate builders one by one from buildbot to swarmbucket
without changing bucket name.

The problem with this approach is that if there is no buildbot master
polling a bucket and someone misspels a builder name, the build will
hang ignored for 48h until expired.

Eventually we decided that we will always change bucket name from
master.x.y to luci.x.y (because it is not a buildbot master), which
means we can immediately mark a build as failed if it specifies a
non-existent builder. Do that in this CL.

R=vadimsh@chromium.org, estaab@chromium.org
BUG=

Change-Id: Iebd74265c51936e4e229a734fd7290bbf973beb2
Reviewed-on: https://chromium-review.googlesource.com/433520
Reviewed-by: Erik Staab <estaab@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>