[monorail] allow project owners to ban users site-wide
I had a hard time shoe-horning this into the permissions model. Is there
a cleaner way to do this? I tried adding a new BAN_USER permission and
adding that to OWNER_ACTIVE_PERMISSIONSET - but when you go to check that
on a user profile page, there is no 'project' artifact to evaluate the
check against. The intent is for any user who is trusted with OWNER role
on *any* project to be able to update the banned status of any other user.

BUG=monorail:2176

Change-Id: Id39e25a8bdaa7d84fedf6c9d43f76c2672f732eb
Reviewed-on: https://chromium-review.googlesource.com/435421
Commit-Queue: Sean McCullough <seanmccullough@chromium.org>
Reviewed-by: Jason Robbins <jrobbins@chromium.org>