buildbucket: implement ACL sets
Most of buckets defined in one file have similar, if not identical, sets
of ACLs. This is verbose.

Implement acl sets in the configuration. An ACL set is a named set of Acl
messages. It can be referenced by a bucket, in which case all of the
acls in the set are inlined in the bucket.

The internal representation of datastore does not change: ACLs are
stored inlined.

R=vadimsh@chromium.org, estaab@chromium.org
BUG=

Change-Id: I3aac0b607ec9b60dd229e5c10f7d0390b0b2f299
Reviewed-on: https://chromium-review.googlesource.com/435638
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>