android-docker: Remove all uses of adb from container scripts.
Turns out I can make the cron run as root too. So now it can add to the
cgroups directly instead of rebooting the devices to trigger the udev
events to do the adding for it... if that makes any sense.

Also log a warning if not root.

BUG=670879

Change-Id: Id53f3c2006543cbb10dcef76cd394d4a3de114a1
Reviewed-on: https://chromium-review.googlesource.com/442677
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Benjamin Pastene <bpastene@chromium.org>