Enable bucket pausing in BuildBucket.
Add an administrative endpoint to pause/resume individual buckets. Pausing
a bucket has the effect of causing "peek" calls to return nothing, even
if there are builds available for lease in the bucket.

Pausing is a harmless way to stop systems from pulling builds from
BuildBucket without disrupting the flow of systems that schedule those
builds. This can be used in the case of an overload to mitigate
BuildBucket's influence.

BUG=None
TEST=unit

Change-Id: Ib47812c59b89c77d558ad157bea32d7cac61b20e
Reviewed-on: https://chromium-review.googlesource.com/443844
Commit-Queue: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>