[findit] Add callback_url and push endpoint to swarming tasks.
This is in preparation for an upcoming change that will make the
pipelines that process swarming tasks asynchronous.

This change should not alter the behavior of the pipelines, but
it will make swarming reach our endpoint, which we'll do nothing
with since we are still not populating the property that connects
a swarming task entity with the pipeline that triggered it.

R=chanli,lijeffrey
BUG=

Change-Id: Ia7c92d4750db5b8417c706c29fd3d08b1fe58ff0
Reviewed-on: https://chromium-review.googlesource.com/446037
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Jeffrey Li <lijeffrey@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>