[Monorail] Add DB support for "niche" custom fields.
Basically there was a request to allow subteams to have custom fields that
do not get in the way of people outside that subteam.  I thought that I had
that covered with the applicable_type DB field, but I actually need an
additional boolean.

This is 1 of 3 CLs that will implement the ability to define custom fields
that require the user to click "Show all fields" to set a value when there
is no previous value for that field on that issue.

BUG=monorail:2297

Change-Id: I1c9dd4f934205cf1d11c157a1476dfdea99fa316
Reviewed-on: https://chromium-review.googlesource.com/446740
Reviewed-by: Aaron Gable <agable@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>