buildbucket: default canary template percentage
If we don't run canary on users' buckets, we won't notice if the canary is
broken. If then we promote canary to prod, it will break all builds in
the user bucket. This is fragile.

Introduce default canary percentage. This also requires monitoring at
lower levels of the stack, e.g. inside kitchen.

R=dnj@chromium.org, vadimsh@chromium.org
BUG=

Change-Id: I35f6ec28425b03cc94f74634475f8d67b9b957e5
Reviewed-on: https://chromium-review.googlesource.com/446661
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Daniel Jacques <dnj@chromium.org>