Add "buildbucket-viewer" GAE application.
Add the "buildbucket-viewer" GAE application. This provides a consistent
framework for a versatile BuildBucket build status viewer. It also
allows open queries.

This app is fully tied to authentication and configuration services. It
allows protobufs to be used to define per-project fixed views. It also
allows open queries to experiment with views. All views/queries have a
fairly versatile set of parameters that can be applied to define the
views.

When possible, click-through links to Milo and BuildBucket are
integrated into the results for easy traversal.

This app is currently experimental and unmaintained. If it is useful, we
may integrate it as a BuildBucket UI or small standalone app sometime in
the future.

TBR=estaab@chromium.org
BUG=None
TEST=None

Change-Id: I2fe518ca073b8dc262328461eb8b7fa53cc51bf2
Reviewed-on: https://chromium-review.googlesource.com/409530
Reviewed-by: Daniel Jacques <dnj@chromium.org>
Commit-Queue: Daniel Jacques <dnj@chromium.org>