kitchen: fetch repo with a recipe
recipes.py remote internally uses GitBackend.checkout() which is
problematic and not trivial to fix, see
https://codereview.chromium.org/2720853002/.
In particular, it does not refetch a branch if it is present locally
(which is NOT what kitchen needs).

The only value `recipes.py remote` provides is to fetch a repository,
parse recipes.cfg file and run recipes.py within the repo. The price we
pay is depending on presence of recipes-py CIPD package and depending on
buggy Python code. The price overweights the value, we could achieve the
same value trivially in Go code. This CL does that.

R=dnj@chromium.org
BUG=

Change-Id: Icee49a885f4ea76c24c4ab3bd8cb7caecefee765
Reviewed-on: https://chromium-review.googlesource.com/448157
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Daniel Jacques <dnj@chromium.org>