Log an error for all failures to chrome-build-extract
Previously we were receiving 404s from Buildbot when build expired, but this
should not happen for chrome-build-extract, because it's backed by Milo and it
does not expire builds ever. This may change in the future, but for now we
should log all failures as an error to get alerts.

Change-Id: I305193c921992a6b0db27fef26559226cf4805be
Reviewed-on: https://chromium-review.googlesource.com/450098
Commit-Queue: Sergiy Byelozyorov <sergiyb@chromium.org>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>