buildbucket: fix permissions check in search
Build search does not perform a check if a requester is searching for a
build by retry_of attribute. It is possible to see builds that are
retries of a build that belongs to a bucket that we don't have access
to.

Fix that and add a test.

R=vadimsh@chromium.org
BUG=

Change-Id: Iffb63c3b2f1be74b7564ed131d84762808f0e1d4
Reviewed-on: https://chromium-review.googlesource.com/452556
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>