Master Manager: Use "varz" instead of "buildstate"
The "buildstate" JSON endpoint is dead, but "varz" lives on. Have master
manager use the "varz" endpoint instead of the "buildstate" one.

To do this, "running_builds" is no longer a list of (builder,
buildnumber) tuples. Now, it is a list of (builder, total_builds). This
results in some trivial code tweaks using sums instead of counts, but
nothing big.

BUG=chromium:701439
TEST=unit
  - Unit tests pass.

Change-Id: Ifb51b157912f9276596f14b62ce48d81ed4b7a8c
Reviewed-on: https://chromium-review.googlesource.com/454877
Commit-Queue: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Ryan Tseng <hinoka@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>