swarmbucket: remove support for CIPD
The current implementation of user-supplied CIPD packages at the
swarmbucket level is incorrect: that packages are installed into the
swarming task rundir and are not exposed to the recipes unless a user
guesses recipe workdir. Also it is possible for a user override system
packages, such as kitchen.

Remove support for user-supplied CIPD packages until this is fixed.

R=vadimsh@chromium.org, phosek@chromium.org
BUG=

Change-Id: I96b392440adaffce0eb80c42ddd02eecba0f1722
Reviewed-on: https://chromium-review.googlesource.com/455338
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Petr Hosek <phosek@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>