[Findit] Fix bug when update try job results with culprit information.
Before we only check if the last one in try job result is the current one,
but in reruns sometime 2 try jobs are running at the same time and that will cause
redundant data in try job result. To prevent that, we need to check the whole list instead.

BUG=702690

Change-Id: I342c7896b84ae6311e7ab6557c626563b7d61a65
Reviewed-on: https://chromium-review.googlesource.com/456590
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Commit-Queue: Chan Li <chanli@chromium.org>