[Findit] Fixing Revert.timestamp not to use timezone
BUG=703502

It seems the timestamp returned from time_util.UTCDatetimeFromNaiveString is in the format datetime.datetime(2017, 3, 19, 1, 28, 20, 351410, tzinfo=<UTC>), with "tzinfo=<UTC>" throwing the exception when calling .put().
It appears the timestamps in Rietveld are already in UTC, making this unnecessary and just DatetimeFromString suffices and is in a format that can be written to the datastore.

Change-Id: I8e7e797109bb709afd9d5325bdc2648cc2228470
Reviewed-on: https://chromium-review.googlesource.com/457294
Commit-Queue: Jeffrey Li <lijeffrey@chromium.org>
Reviewed-by: Roberto Carrillo <robertocn@chromium.org>