buildbucket: do not mark swarmbucket build as STARTED
In the past Swarming did not send a PubSub notification when task
started, and buildbucket had to mark builds as STARTED immediately. This
is not ideal because users, including CQ, think that a build has
started, although it did not.

Now Swarming sends a PubSub notification on any change to "state" field,
so we don't have to mark a build as STARTED immediatelly.

R=vadimsh@chromium.org, estaab@chromium.org
BUG=

Change-Id: Id966936d9e4ad4c9db5bef8cdd61ef8d10515e2a
Reviewed-on: https://chromium-review.googlesource.com/457121
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>