buildbucket: refactor build creation
Change service.add_many_async from a giant multi-phase function to a function
with a small body but a few sub-functions

Merge _put_build into add_many_async and create swarming tasks before updating
tag indexes. This means:

* if swarming task creation fails and a client retries results, we won't be
  adding invalid entries to the tag index.
* swarming.create_task_async can add indexed tags to the build and they will be
  added to the tag index.

Remove swarming.is_for_swarming. A build is for swarming if its bucket is 
for swarming. In swarming.create_task_async raise an error if a build is for
a non-swarming bucket, or if task template is not defined for this instance
of buildbucket.

R=vadimsh@chromium.org
BUG=

Change-Id: If55e1c27686834c2672b1e3d07b1956c0e90ee25
Reviewed-on: https://chromium-review.googlesource.com/461239
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>