Update infra "luci_deploy" configuration (#2)
Update the "luci_deploy" configuration and associated scripts to be
simpler and more robust against future changes.

Currently, "luci-deploy" runs "bootstrap/luci-deploy-init.py", which is
a custom script that sets up bootstrapping by hackily using internal
functions in the Python and Go bootstrap code. This is currently broken,
since that code moved without updating "luci-deploy-init.py".

Rather than fight this eternal fight, update the "luci_deploy"
configuration to directly use the Python and Go bootstrap scripts. This
leverages a new "script" protobuf layout in "luci_deploy" that is
versatile enough to enable this.

By default, the Go toolchain is installed into <infra>/../goroot. This
is not acceptable to "luci_deploy", since it assumes ownership of the
parent directory of its checkouts. In order to use the Go bootstrap's
"env.py", we need to modify it to add flags allowing it to run against a
"goroot" in a different directory. We choose
"<infra>/.luci_deploy/goroot".

BUG=chromium:706269
TEST=local
  - Ran upstream "luci_deploy" with this new configuration, seems to
    work.

Change-Id: I91464b4c2e83e4c89dc8130da0b5fd498d7d9a8a
Reviewed-on: https://chromium-review.googlesource.com/462441
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Daniel Jacques <dnj@chromium.org>