Kitchen: Add "-prefix-path-env" flag.
Add "-prefix-path-env" flag, which instructs Kitchen to prefix a series
of path elements to the PATH environment variable. This will be used to
establish a system PATH that can be used by both Kitchen and its
child processes.

This also includes some formatting and cleanup of Kitchen's flag
definitions for easier readability and traversability.

BUG=chromium:703868
TEST=None

Change-Id: I36fac724f2fa49a6d39c599dc389250e8753bfae
Reviewed-on: https://chromium-review.googlesource.com/468546
Commit-Queue: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>