Return a deepcopy of MetricFieldsValues.
Some of the values are Distribution objects - it's critical that we make deep
copies of these, otherwise their internals can be modified while we're
serialising the metrics protos, causing a race.

Sanity check that the bucket is in the range that will be included in the serialised metric proto.

Bug: 709909
Change-Id: I12e98fd8c12e93285ded0b0378d00b90ee021b18
Reviewed-on: https://chromium-review.googlesource.com/472650
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Sergey Berezin <sergeyberezin@chromium.org>
Commit-Queue: Dave Sansome <dsansome@chromium.org>