Git: honor "-C" when determining clone directory.
Currently, the Git wrapper will parse the "git clone" flags to determine
the cloned destination directory so that it can destroy it in the event
of a retry. However, the "clone" target directory is made relative to
Git's working directory, which can be affected by the use of the "-C"
flag.

Update flagparse to store its parsed flags in a map, so we can easily
determine a flag's value.

Update Git wrapper to use the contents of the "-C" flag to determine the
current working directory, and apply that to the "clone" target
directory.

Finally, add a test for using "clone" with "-C".

BUG=None
TEST=unit

Change-Id: I66c7ffddead9f10ede502e2a0dd0cabe7a3f3d55
Reviewed-on: https://chromium-review.googlesource.com/479459
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Daniel Jacques <dnj@chromium.org>