third_party_packages recipe: package Git and Python
We use Git and Python through the fleet, but their deployment to build machines
is not controlled. Different machines have different versions. Some are old
which is a security risk. We must update entire fleet when a security patch is
released. On Windows we deploy Python via depot_tools.

The purpose of this new recipe is for each third party software that we care
about, check the latest version, check if we have a CIPD package for it, and if
not, build it and upload the package. The initial version covers only
- Git for Linux, Mac and Windows (32bit)
  CIPD prefix is "infra/git/"
- Python for Linux and Mac
  CIPD prefix is "infra/python/cpython/"

This recipe is meant to be run periodically, for example once a day.

R=iannucci@chromium.org, vadimsh@chromium.org
BUG=714852
Change-Id: I58b6be3b283ad7cd8b40bb3f300a83d540da8631

Change-Id: I58b6be3b283ad7cd8b40bb3f300a83d540da8631
Reviewed-on: https://chromium-review.googlesource.com/483944
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>