[Monorail] Use status 400 for user queries that exceed fulltext search limit.
This addresses the case where the user submitted a huge fulltext
query, longer than 2000 characters.

With or without this CL, the user will see a search result page with
no results and an error message that says "10 search backends did not
respond or had errors. These results are probably incomplete."  The
difference with this CL is that the backends will respond with status
code 400 rather than 500, so it will not trigger a monitoring alert.

Bug: monorail:2624
Change-Id: Ie51f87d1650c9d3cfba5e99e44332480e963edd6
Reviewed-on: https://chromium-review.googlesource.com/494091
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>