[Monorail] Autolink crbug.com references.
If the user types text into a comment like "crbug.com/12345", it will
be automatically linked to that issue in monorail, and a tooltip will
be shown with the issue summary, and the link will be crossed out if
the issue is closed.

This CL registers a new autolink handler and refactors some of the
existing code.  The autolinking code is still pretty complicated, but
at least now the MonorailRequest does not get passed through as many
functions.

Bug: monorail:1033
Change-Id: Ib13282cd022160c23da45baa3a1b83ff14561c05
Reviewed-on: https://chromium-review.googlesource.com/493830
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Aaron Gable <agable@chromium.org>