[SECOND RELAND][Findit] Get step log for swarmbucket jobs.
The issue with the original patch is that it used a real http client for
testing instead of a fake one.

And it had to be reverted once again because one of the tests was still
making http calls to buildbucket
(swarming_util_test.testGetStepMetadataStepNone)

This time, I disconnected my network cable to run the tests to check
that it's no longer using the network.

CL Containing the last revert:
https://chromium-review.googlesource.com/c/495866

This subsubmes: https://chromium-review.googlesource.com/c/495588/

R=chanli,stgao,lijeffrey

Bug:

Change-Id: I2b4742fbf108c561e0b7541fd4e9521cf56db3aa
Reviewed-on: https://chromium-review.googlesource.com/495667
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>