buildbucket: shorten returned API Explorer link
In practice, the Swarming API Explorer URLs returned by get_task_def API are so
long, they cannot be pasted in Google Hangouts. Shorten them.

Add a parameter to avoid generation of the link in cases when it is not needed,
e.g. try-recipes

R=iannucci@chromium.org

Change-Id: I4005acaa2a9dcd42bc0b1d2690f5b2bbcd3edf78
Reviewed-on: https://chromium-review.googlesource.com/496790
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>